cleanup - explicit application configuration
