--- api/src/main/kotlin/github/com/hukuta94/delivery/api/startup/configuration/adapter/kafka/KafkaConfiguration.kt
@@ -1,13 +1,0 @@
-package github.com.hukuta94.delivery.api.startup.configuration.adapter.kafka
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.Import
-import org.springframework.kafka.annotation.EnableKafka
-
-@Configuration
-@EnableKafka
-@Import(
-    KafkaConsumerConfiguration::class,
-    KafkaProducerConfiguration::class,
-)
-open class KafkaConfiguration
