+++ api/src/main/kotlin/github/com/hukuta94/delivery/startup/DeliveryApplication.kt
@@ -1,0 +1,8 @@
+package github.com.hukuta94.delivery.startup
+
+import github.com.hukuta94.delivery.startup.configuration.DeliveryApplicationConfiguration
+import org.springframework.boot.runApplication
+
+fun main(args: Array<String>) {
+    runApplication<DeliveryApplicationConfiguration>(*args)
+}
