+++ api/src/main/kotlin/github/com/hukuta94/delivery/startup/configuration/core/application/port/PortConfiguration.kt
@@ -1,0 +1,20 @@
+package github.com.hukuta94.delivery.startup.configuration.core.application.port
+
+import github.com.hukuta94.delivery.infrastructure.adapter.grpc.GeoServiceGrpcClient
+import github.com.hukuta94.delivery.startup.configuration.infrastructure.adapter.grpc.GrpcConfiguration
+import io.grpc.ManagedChannel
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Import
+
+@Configuration
+@Import(
+    GrpcConfiguration::class,
+)
+open class PortConfiguration {
+
+    @Bean
+    open fun getLocationPort(
+        grpcChannel: ManagedChannel,
+    ) = GeoServiceGrpcClient(grpcChannel)
+}
