+++ api/src/main/kotlin/github/com/hukuta94/delivery/startup/configuration/core/application/usecase/OrderUseCaseConfiguration.kt
@@ -1,0 +1,39 @@
+package github.com.hukuta94.delivery.startup.configuration.core.application.usecase
+
+import github.com.hukuta94.delivery.core.application.usecase.order.AssignCourierToOrderUseCase
+import github.com.hukuta94.delivery.core.application.usecase.order.CreateOrderUseCase
+import github.com.hukuta94.delivery.core.application.usecase.order.impl.AssignCourierToOrderUseCaseImpl
+import github.com.hukuta94.delivery.core.application.usecase.order.impl.CreateOrderUseCaseImpl
+import github.com.hukuta94.delivery.core.domain.service.DispatchService
+import github.com.hukuta94.delivery.core.port.repository.domain.CourierRepositoryPort
+import github.com.hukuta94.delivery.core.port.GetLocationPort
+import github.com.hukuta94.delivery.core.port.repository.domain.OrderRepositoryPort
+import github.com.hukuta94.delivery.core.port.repository.UnitOfWorkPort
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+open class OrderUseCaseConfiguration {
+
+    @Bean
+    open fun createOrderUseCase(
+        orderRepository: OrderRepositoryPort,
+        getLocationPort: GetLocationPort,
+    ): CreateOrderUseCase = CreateOrderUseCaseImpl(
+        orderRepository = orderRepository,
+        getLocationPort = getLocationPort,
+    )
+
+    @Bean
+    open fun AssignCourierToOrderUseCase(
+        orderRepository: OrderRepositoryPort,
+        courierRepository: CourierRepositoryPort,
+        dispatchService: DispatchService,
+        unitOfWork: UnitOfWorkPort,
+    ): AssignCourierToOrderUseCase = AssignCourierToOrderUseCaseImpl(
+        orderRepository = orderRepository,
+        courierRepository = courierRepository,
+        dispatchService = dispatchService,
+        unitOfWork = unitOfWork,
+    )
+}
